fix: Fix add-contact page being unusable with smaller viewports
If you e.g. have the Android keyboard up (especially with the phone in
landscape mode), widgets lower on the page end up hidden behind the
keyboard.

This change makes it so that scrolling is allowed on the page, and so
e.g. hitting enter when you finish filling in a textfield will scroll to
the next one if it's out-of-view.